+++ app/views/degreecourses/index.blade.php
@@ -1,3 +1,5 @@
+@extends('layouts.main')
+
 @section('scripts')
 	@include('layouts.partials.datatables-script')
 	<script type="text/javascript" class="init">
@@ -18,13 +20,10 @@
 	</ol>
 @stop
 @section('main')
-	<!-- <h3>Studiengangsmanagement
-		<button class="btn btn-success btn-xs" data-toggle="modal" data-target="#myModal">
-	  		<span class="glyphicon glyphicon-plus"></span>
-		</button>
-	</h3> -->
 	
-	@include('layouts.partials.add-button-modal', ['buttonLabel' => 'Studiengang hinzufügen'])
+	@if ($currentUser->hasRole('Admin') OR $currentUser->can('add_degreecourse'))
+		@include('layouts.partials.add-button-modal', ['buttonLabel' => 'Studiengang hinzufügen'])
+	@endif
 	
 	<div class="row">
 		<div class="col-sm-12" style="margin-bottom: 5px;">
@@ -43,15 +42,21 @@
 						@foreach( $degreecourses as $degreecourse )
 							<tr>
 								<td>{{ $listofdegrees[$degreecourse->degree_id] }}</td>
-								<td>{{ link_to_route('showDegreecourse_path', $degreecourse->name, $degreecourse->id) }}</td>
+								<td>
+									@if ($currentUser->hasRole('Admin') OR $currentUser->can('edit_degreecourse'))
+										{{ link_to_route('editDegreecourseInformation_path', $degreecourse->name, $degreecourse->id) }}
+									@else
+										{{ $degreecourse->name }}
+									@endif
+								</td>
 								<td>{{ $degreecourse->short }}</td>
 								<td>{{ $listofdepartments[$degreecourse->department_id] }}</td>
 								<td>
-									{{ Form::open(array('class' => 'inline', 'method' => 'DELETE', 'route' => array('deleteDegreecourse_path', $degreecourse->id))) }}
-										{{ HTML::decode(link_to_route('showDegreecourse_path', '<i class="glyphicon glyphicon-edit"></i>', array($degreecourse->id), array('class' => 'btn btn-xs btn-warning'))) }}
-
-										{{ Form::button('<i class="glyphicon glyphicon-remove"></i>', array('type' => 'button', 'class' => 'btn btn-xs btn-danger', 'data-toggle' => 'modal', 'data-target' => '#confirmDelete', 'data-title' => 'Studiengang löschen', 'data-message' => 'Wollen Sie den Studiengang wirklich löschen?')) }}
+									@if ($currentUser->hasRole('Admin') OR $currentUser->can('delete_degreecourse'))
+										{{ Form::open(array('class' => 'inline', 'method' => 'DELETE', 'route' => array('deleteDegreecourse_path', $degreecourse->id))) }}
+											{{ Form::button('<i class="glyphicon glyphicon-remove"></i>', array('type' => 'button', 'class' => 'btn btn-xs btn-danger', 'data-toggle' => 'modal', 'data-target' => '#confirmDelete', 'data-title' => 'Studiengang löschen', 'data-message' => 'Wollen Sie den Studiengang wirklich löschen?')) }}
 										{{ Form::close() }}
+									@endif
 		    					</td>
 							</tr>
 						@endforeach
@@ -62,7 +67,7 @@
 	</div>
 
 	<!-- add form modal dialog -->
-	@include('degree_courses.partials.add-form')
+	@include('degreecourses.partials.add-form')
 
 	<!-- delete confirmation modal dialog -->
 	@include('layouts.partials.delete_confirm')
